Improve README formatting
I found the README to be very well written, but the formatting made it difficult to read. I've taken some time to adjust the formatting to make headings, sections, and important information stand out more. I think the result is much easier to scan while introducing almost no changes to the content itself.